delete button works

+++ src/main/java/com/codeup/adlister/dao/Ads.java
@@ -16,5 +16,5 @@ public interface Ads {
 
     void editAd(Ad ad);
 
-    void deleteAd(int id);
+    void deleteAd(long id);
 }
